refactor:SpeakerSelect와대화비율그래프 하나의 컴포넌트로 통합& speakerselect dashbordHead로 통일
