+++ src/components/molecules/dashboard/SpeakerSelect.tsx
@@ -13,10 +13,10 @@ const SpeakerSelectBox = styled(FlexColumnDiv)<{ alignItems?: string }>`
   font-size: 12px;
 `;
 
-const CurrentSpeaker = styled(Span)`
-  color: #7e848a;
-  margin-bottom: 15px;
-`;
+// const CurrentSpeaker = styled(Span)`
+//   color: #7e848a;
+//   margin-bottom: 15px;
+// `;
 
 const SelectNotice = styled(Span)`
   font-size: 15px;
@@ -58,7 +58,7 @@ const SpeakerSelect: React.FC<SpeakerSelectProps> = ({ alignItems }) => {
 
   return (
     <SpeakerSelectBox alignItems={alignItems}>
-      <CurrentSpeaker>강조할 대화자</CurrentSpeaker>
+      {/* <CurrentSpeaker>강조할 대화자</CurrentSpeaker> */}
       <Select
         value={selectedSpeakerIndex === -1 ? "전체" : selectedSpeakerIndex}
         onChange={handleChangeSpeaker}
